fix(logger) new module
